load manuscript card details lazily on expand
